migration refactored and conditiona applied for good

+++ database/migrations/2023_01_05_130106_update_general_settings_from_horizontal_to_vertical.php
@@ -13,13 +13,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::dropIfExists('general_settings');
-        Schema::create('general_settings', function (Blueprint $table) {
-            $table->id('id');
-            $table->string('key');
-            $table->text('value')->nullable();
-            $table->foreignId('branch_id')->nullable()->references('id')->on('branches')->onDelete('CASCADE');
-        });
+        // Schema::dropIfExists('general_settings');
+        if(! Schema::hasTable('general_settings')) {
+            Schema::create('general_settings', function (Blueprint $table) {
+                $table->id('id');
+                $table->string('key');
+                $table->text('value')->nullable();
+                $table->foreignId('branch_id')->nullable()->references('id')->on('branches')->onDelete('CASCADE');
+            });
+        }
     }
 
     /**
